rename eventListView
