+++ lib/components/event_list_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-class EventListView extends StatelessWidget {
+class MyEventListView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ListView.builder(
